add pause and new track

+++ 3D Racing 25/Assets/Scripts/Settings/GraphicsQualitySetting.cs
@@ -46,6 +46,6 @@ public class GraphicsQualitySetting : Setting
     }
     private void Save()
     {
-        PlayerPrefs.SetFloat(title, currentLevelIndex);
+        PlayerPrefs.SetInt(title, currentLevelIndex);
     }
 }
