+++ 3D Racing 25/Assets/Scripts/Settings/ResolutionSetting.cs
@@ -55,6 +55,6 @@ public class ResolutionSetting : Setting
     }
     private void Save()
     {
-        PlayerPrefs.SetFloat(title, currentResolutionIndex);
+        PlayerPrefs.SetInt(title, currentResolutionIndex);
     }
 }
